chore(update-server): refactor current behavior into a submodule
This moves the balena updater into a submodule of update-server, in preparation for adding a buildroot submodule
