+++ update-server/otupdate/balena/selftest.py
